Refactoring, got output planes to work.
Spent a lot of time trying to work out why the views for the indirect
draw params and the output planes were getting swapped in
renderdoc (like, showing eachother's data).

I tore out the shader binding points and replaced them with manual
bindings, but that didn't work.

I ended up just reordering everything to be issued in the ordering
that renderdoc seemed to be trying to display and that did it.

What a waste of an evening.